Update: to be improved running CI under more real circumstances
Since
https://github.com/nix-community/home-manager/commit/6a1922568337e7cf21175213d3aafd1ac79c9a2e,
it is not possible to activate another user home directory.

Under the GH CI environment, the real user executing the activation
script is `runner`, whereas we are trying to set other users
directories here.

It can be worked around by adding a new user, and running the
activation script with that user, but is fragile.